Changed all REAL variables and constants to have KIND set by prec_def. Changed error reporting in wfnho.f90 to use module bexception.
Still need to change some conditionals that check for errors by making them relative to the precision used.

+++ src/time.f90
@@ -25,7 +25,7 @@ MODULE time
   USE prec_def
 
   INTEGER     :: it     !< current iteration of time
-  REAL*8      :: t      !< current time during time evolution [fm/c]
+  REAL (Long) :: t      !< current time during time evolution [fm/c]
   INTEGER     :: Nt     !< number of timesteps in current mode (adiabatic or
                         !!  time)
